chore: clean up token retreival process
