+++ lib/spotify.js
@@ -1,3 +1,5 @@
+import { getTokens } from "../utils/auth";
+
 /**
  * Gets fetch request init options
  * @param {Object} data
@@ -6,7 +8,7 @@
  */
 export async function getRequestInitOptions(data, method, tokens) {
   if (!tokens.current.access_token) {
-    const response = await fetch("/api/v1/tokens");
+    const response = await getTokens();
     const json = await response.json();
     tokens.current.access_token = json.access_token;
     tokens.current.refresh_token = json.refresh_token;
